Move rviz related code into a separate file
